Add parallel_fsps.py to demonstrate parallel FSPS
This uses ipython cluster to send jobs to several nodes. The advantage
of using ipython cluster is that you can instantiate python-fsps only
once on each node, and then repeatedly call that instance so that
overhead from loading isochrones and spectra is minimized. See the
demo/parallel_fsps.py docs for details on how to run.